+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [0.0.9] - 2023-06-27
 ### Fix
 - Format logs before calling log macro
+- Don't fail when backing up and restoring plugins without docker images
 
 ## [0.0.8] - 2023-06-26
 ### Fix